Refactor: extract util package
