+++ src/main/groovy/org/hidetake/gradle/swagger/generator/ValidateSwagger.groovy
@@ -3,11 +3,9 @@ package org.hidetake.gradle.swagger.generator
 import com.github.fge.jsonschema.main.JsonSchemaFactory
 import groovy.util.logging.Slf4j
 import org.gradle.api.DefaultTask
-import org.gradle.api.tasks.CacheableTask
-import org.gradle.api.tasks.InputFiles
-import org.gradle.api.tasks.OutputFile
-import org.gradle.api.tasks.SkipWhenEmpty
-import org.gradle.api.tasks.TaskAction
+import org.gradle.api.tasks.*
+import org.hidetake.gradle.swagger.generator.util.Mappers
+import org.hidetake.gradle.swagger.generator.util.Resources
 
 @Slf4j
 @CacheableTask
